refactor: TechNews 커스텀 훅 분리
테크뉴스 썸네일 슬라이드 기능을 useTechNewsSlider.tsx로 분리
